before:
```
▶ time dune exec -- ./bench/irmin-pack/layers.exe -n 300
4500 commits completed in 30.63s.
[0.007s per commit, 147 commits per second]
dune exec -- ./bench/irmin-pack/layers.exe -n 300  24.61s user 6.61s system 97% cpu 32.100 total
```

after:
```
▶ time dune exec -- ./bench/irmin-pack/layers.exe -n 300
4500 commits completed in 27.53s.
[0.006s per commit, 163 commits per second]
dune exec -- ./bench/irmin-pack/layers.exe -n 300  21.97s user 6.12s system 99% cpu 28.245 total
```